### 2022-04-28 #4359 @editorialbot
[PRE REVIEW]: islatu: A Python package for the reduction of reflectometry data
**Submitting author:** @RBrearton (<a href="http://orcid.org/0000-0002-8204-3674">Richard Brearton</a>)
**Repository:** https://github.com/RBrearton/islatu
**Branch with paper.md** (empty if default branch): islatu_paper
**Version:** v1.0.0
