### 2022-05-07 #4359 @jgostick
Hello @andyfaff.  I have tried to find an email address to contact you directly, but you've done a great job at keeping it secret ;-)  Anyway, hopefully you get notified by Github.  I would like to ask if you're willing and able to review this software submission.  The process of reviewing in JOSS is a bit different than a *normal* journal.  There is a checklist of items for you to look for, and you work with the authors to get their package into publishable shape, rather than an accept/reject decision.  It's also all done publicly on a github thread.  Please let me know via response here, or via email (jgostick at gmail) if you're able to do this, or perhaps to recommend alternative reviewers.  Thanks, Jeff.
