### 2022-05-16 #4397 @editorialbot
[REVIEW]: islatu: A Python package for the reduction of reflectometry data
**Submitting author:** @RBrearton (<a href="http://orcid.org/0000-0002-8204-3674">Richard Brearton</a>)
**Repository:** https://github.com/RBrearton/islatu
**Branch with paper.md** (empty if default branch): islatu_paper
**Version:** 1.0.7
