### 2022-05-16 #4397 @andyfaff
License: yes, but the setup.py file needs updating with the correct licence identifier.

Quality of writing: Good, but there are a few spelling mistakes that could be picked up by a spell checker.

State of the field: It might be worth mentioning a couple of other reduction packages in the arena, and how islatu works compared to them.

Community guidelines: there aren't any that I can discern.

Installation instructions: the instructions in README.md don't work out of the box, they should be updated. There is a requirements.txt file that works.

Installation: does not proceed according to the instructions.

References: It might be worthwhile including links to the articles cited?
