### 2022-05-18 #4397 @jgostick
Hi @daguiam
You'll want to generate an 'editorial checklist' for yourself, so you can see which aspects of the codebase to inspect and approve.  You can do this by typing ``@editorialbot generate my checklist`` into a comment box, then you'll have a list like @andyfaff's above.  The review process typically takes at least a month, so if you can set aside an afternoon or two of the next few weeks, then that should be sufficient, at least for the first pass of reviews.
