### 2022-07-01 #4397 @daguiam
Hi! I have continued the review since the last issues were resolved and I have some comments below. Overall, despite some bugs that I have added as issues, the package meets most of the checks.

_Functionality:_
The issues with the installation steps have been corrected and it is now straightforward to install the package.
Following through the provided notebooks the functionality appears to work as claimed in terms of reducing the reflectometry data with appropriate signal processing, error propagation and data correction.
The packaged has some hardcoded strings related to paths and where to find the datasets (with hardcoded network paths to the Diamond I believe) if the Profile is not provided the direct data folder path, assuming that the user is inside the network and has access to those directories. These could be removed or pass an argument flag to look for the datasets in those directories.

_Documentation:_
Example usage: The documentation provided is sufficient to follow the workflow of I07 reflectometry data reduction using the notebooks. The `process_xrr.py` CLI is not documented in the readthedocs.
Statement of need: The statement of need is better described in the paper than in the software documentation.
Missing community guidelines on how to contribute to the project or contact the authors.
The `README.md` could provide a direct link to the documentation website and the notebook that gives an example of the data reduction workflow.

_Software paper_
The software paper describes well the need for such a package and who the target audience is. It also presents the overall functionality of the package. It is lacking a comparison with other reflectometry data reduction packages or efforts made, as stated by @andyaff. The quality of the writing is good and reference to cited packages are provided.
