### 2022-07-25 #4397 @RBrearton
Hi @daguiam,

With regards to hardcoded paths, I think these are only given where appropriate at the moment. The only diamond filesystem path in the repo is in a default argument to `i07reduce` in runner.py, which seems like a reasonable default since that function is designed to specifically handle data from diamond's i07 beamline. Another local path can of course be passed in place of that default argument.

The CLI also often checks for a directory called 'processing', but these checks should all be harmless and no errors should be raised if the directory doesn't exist. I'm aware that there was such an error, but that has now been fixed!

The documentation has been updated to include contact details, a brief contributing section and a discussion (with examples) on how to use the command line interface. These examples use the diamond directory structure, but changing the data directory to be a local directory should be completely clear. I just chose this because the people most likely to come across the readthedocs page are diamond users, and this should make more sense to them.

I updated the readme.md to link directly to the docs, as suggested. I'll provide another update when I update the paper to include references/comparisons to other reflectivity reduction tools.
