### 2022-08-08 #4397 @jgostick
Comments on your paper:
- [ ] It seems like the summary section ends abruptly. It feels like it needs a closing sentence like. "For the above reasons all reflectometry experiments require intensive data analysis and adjustments, which are cumbersome to implement and error prone. Reliable and rigorously tests numerical packages are essential...blah, blah, blah..."
- [ ] "simplify the handling of errors" should probably be "handling of uncertainties" or "experimental errors"?
- [ ] Is it possible to put a screen shot or some sort of image in the paper?  It's just pure text at the moment.  This is only a suggestion, but what about showing the initial and final reflectivity curves that you have in your [example here](https://islatu.readthedocs.io/en/latest/i07_reflectivity.html)
- [ ] Your "statement of need" section points out why someone might need software to do this analysis but should also or alternatively focus on why islatua needed to be written.
- [ ] Are there really no other open source codes?  There is no mention of refnx in the paper, which seems odd.  Perhaps this is why @andyfaff has left the ``Statement of need`` checkbox open?
- [ ] The paper mentions a few things about performance with regard to error propagation ("such that the number of mathematical operations required to propagate the errors is minimized").  Can you elaborate on this at all?
- [ ] Overall the paper is quite short.  I know that JOSS explicitly states that it's about the software not the paper, but this is a chance for you to create a permanent record describing the software. You maybe want to take this opportunity to pad the paper a little bit?  Again this is optional.
