### 2022-08-11 #4397 @andyfaff
I don't think refnx has to be cited, but it would be good to mention other reduction codes apart from islatu, even if it's e.g. something like reductus.
