### 2022-08-19 #4397 @RBrearton
Hi @jgostick, just had some time to go through this.

> It seems like the summary section ends abruptly. It feels like it needs a closing sentence like. ...

Yep, true. Added a closing sentence which also hopefully clarifies where islatu lives in the reflectivity software stack.

>  "simplify the handling of errors" should probably be "handling of uncertainties" or "experimental errors"?

Yep, changed it to "uncertainties" and it's much clearer.

> Is it possible to put a screen shot or some sort of image in the paper?

Yeah... good point. I was lazy with this because the documentation is (hopefully!) easy to find, and technical details can be found there.

> Your "statement of need" section points out why someone might need software to do this analysis but should also or alternatively focus on why islatu needed to be written.

Added a final paragraph to the statement of need section which hopefully satisfies this requirement.

> Are there really no other open source codes? There is no mention of refnx in the paper, which seems odd. Perhaps this is why @andyfaff has left the Statement of need checkbox open?

There's really not much. refnx is, primarily, a fitting program that one would use to analyse a reflectivity curve. The problem is that this isn't what your instrument gives you. Islatu was born out of a need to map large numbers of pixels with meaningless intensities to physically meaningful reflectivity curves, which a user can then plug into a tool like refnx.

Integrating with something like reductus would be awkward, because the philosophy of the program is different. Reductus is a really nice graphical way to reduce (currently neutron) data. Islatu is supposed to be small and specialised – only as large as is required to convert meaningless pixels into a meaningful curve. At that point, you can use your favourite reflectivity fitting program to extract physically meaningful information from your measured data.

I hope this makes the wording in the paper make more sense! I also hope that this round of changes to the paper make this clearer.

>  The paper mentions a few things about performance with regard to error propagation ("such that the number of mathematical operations required to propagate the errors is minimised"). Can you elaborate on this at all?

Well, as with all things, it's possible to write code that does the job but is unbelievably slow. I'm just trying to emphasise here that there was an eye on performance when writing the code. For example, initially we used https://pypi.org/project/uncertainties/ for error propagation but it was sickeningly slow (like 2 orders of magnitude slower than it is now).

Other than that, the phrase about minimising mathematical operations is correct. Wherever possible, tricks with poisson stats were played to decrease the number of times that errors needed to be calculated (e.g. if it's possible to skip calculating errors at some intermediate point in the calculation and compensate for it later, then this will be done). I don't mean that some incredible novel algorithm has been implemented, the point is more to emphasise that thought has been put into making this reasonably quick.

> Overall the paper is quite short. I know that JOSS explicitly states that it's about the software not the paper, but this is a chance for you to create a permanent record describing the software. You maybe want to take this opportunity to pad the paper a little bit? Again this is optional.

If it's possible, I'd prefer to pass on this. Because this is a permanent record, I don't want to pad the paper with specific details that might change in the future. I personally like the idea of something short, sweet and fairly permissive with regards to future changes. I'd like to think that, right now, a reader would get the impression that this is a relatively simple library that maps raw data to R(Q) reasonably quickly; this is exactly what I'd like to get across!

Do let me know what you think of the changes, and what else needs to be done.
