### 2022-08-19 #4397 @jgostick
The paper looks great, the choice of before-after data is very convincing!

Regarding performance, @daguiam had not check this box, but between your comments and @andyfaff's endorsement, I think I am satisfied

Regarding citing relevant softare, I see that you've added references to several existing packages, so I considered @andyfaff's unchecked box on this issue to be satisfied as well.

Overall, I think we're ready to go!  There are a few additional steps though, such as declare a final version number, and minting a doi for the code itself.  I will put all this in the form of a checklist below.
