### 2022-08-19 #4397 @jgostick
- [ ] Double check authors and affiliations on the paper (including ORCIDs)
- [ ] Make a release of the software with the latest changes from the review and report the version number here. This is the version that will be used in the JOSS paper.

- [ ] Set article DOI with ``@editorialbot set <DOI here> as archive``
- [ ] Do a 'dry run' of acceptance with ``@editorialbot recommend-accept``
