### 2022-09-07 #4397 @RBrearton
Hi @jgostick, sorry for the slow response, I've been delayed by the fact that Zenodo is broken atm (it isn't possible to upload files).

At first I thought this was just me being an idiot, but they did post on twitter on the 1st of September that file uploads aren't working, so I suppose it isn't just me.

When Zenodo is back up, I'm ready to make a release number 1.0.7 which will sync with pypi etc. including all changes made in response to this review.
