### 2022-09-22 #4397 @RBrearton
Zenodo is still broken for automatically archiving repositories (at least on my end it is), but I did a manual upload so that we aren't just waiting forever.

I've updated the paper to use the new doi (10.5281/zenodo.7105217). I double checked affiliations and made a new release with all the changes (in fact, I made a lot of released with the changes while trying to get zenodo to work, so we ended on v1.0.7).

Let me know if anything else needs to be done!
