### 2022-09-22 #4397 @RBrearton
Is this a hard requirement? Tim supervised the project. He provided direction and helped both McCluskey and me a great deal. He didn't directly push any commits to the repository, though, so I thought it was appropriate to list him an author on the article but not as an author of the source code.

I can change this if required, but I thought this was a good way of representing authorship!
