### 2022-09-22 #4397 @danielskatz
> Is this a hard requirement? Tim supervised the project. He provided direction and helped both McCluskey and me a great deal. He didn't directly push any commits to the repository, though, so I thought it was appropriate to list him an author on the article but not as an author of the source code.
>
> I can change this if required, but I thought this was a good way of representing authorship!

This is fine in this case. We might suggest that Tim also be added as an author of the zenodo repo based on his role in the project (project contributions don't have to be commits to source code), and then the authors would be consistent, but it's up to you.
