### 2022-09-22 #4397 @jgostick
Hi @RBrearton , thanks to the editorialbot I just noticed that only 1 of your references has a doi. Can update the paper to include these for each reference?
