### 2022-09-23 #4397 @RBrearton
Hi there, I've just updated the DOI's (the Rigaku paper doesn't have a DOI). I've also added Tim as an author on the Zenodo repository (@danielskatz is right about authorship; the place for tracking commits to source code is github, not Zenodo).
